+++ spec/views/arquivos/edit.html.erb_spec.rb
@@ -1,17 +1,26 @@
 require 'rails_helper'
 
 RSpec.describe 'arquivos/edit', type: :view do
-  before do
-    @arquivo = assign(:arquivo, Arquivo.create!(
-                                  sha1: 'MyString'
-                                ))
+  let(:first_file) do
+    path = Rails.root.join('tmp/f1.txt')
+    File.open(path, 'w') { |f| f.write Time.now.to_f }
+    File.new(path)
   end
+
+  let(:arquive) do
+    arquivo = build(:arquivo)
+    arquivo.documento.attach(io: first_file, filename: File.basename(first_file))
+    arquivo.sha1_from_tempfile(first_file)
+    arquivo.save
+    arquivo
+  end
+
+  before { assign(:arquivo, arquive) }
 
   it 'renders the edit arquivo form' do
     render
-
-    assert_select 'form[action=?][method=?]', arquivo_path(@arquivo), 'post' do
-      assert_select 'input[name=?]', 'arquivo[sha1]'
+    assert_select 'form[action=?][method=?]', arquivo_path(arquive), 'post' do
+      assert_select 'input[name=?]', 'arquivo[documento]'
     end
   end
 end