+++ db/migrate/20200820080438_add_idex_to_column_sha1_from_arquivo.rb
@@ -1,0 +1,5 @@
+class AddIdexToColumnSha1FromArquivo < ActiveRecord::Migration[6.0]
+  def change
+    add_index :arquivos, :sha1, unique: true
+  end
+end