+++ db/migrate/20200820082220_create_beneficiarios.rb
@@ -1,0 +1,9 @@
+class CreateBeneficiarios < ActiveRecord::Migration[6.0]
+  def change
+    create_table :beneficiarios do |t|
+      t.string :cpf
+
+      t.timestamps
+    end
+  end
+end