+++ spec/factories/beneficiarios.rb
@@ -1,0 +1,5 @@
+FactoryBot.define do
+  factory :beneficiario do
+    cpf { 'MyString' }
+  end
+end